+++ packages/marko/src/runtime/html/helpers/data-marko.js
@@ -1,11 +1,12 @@
 "use strict";
 
-var escapeXml = require("./escape-xml");
-var escapeSingleQuotes = escapeXml.s;
-var escapeDoubleQuotes = escapeXml.d;
+var escapeQuoteHelpers = require("./escape-quotes");
+var escapeSingleQuotes = escapeQuoteHelpers.___escapeSingleQuotes;
+var escapeDoubleQuotes = escapeQuoteHelpers.___escapeDoubleQuotes;
 var FLAG_WILL_RERENDER_IN_BROWSER = 1;
-// var FLAG_HAS_BODY_EL = 2;
-// var FLAG_HAS_HEAD_EL = 4;
+// var FLAG_HAS_RENDER_BODY = 2;
+// var FLAG_IS_LEGACY = 4;
+// var FLAG_OLD_HYDRATE_NO_CREATE = 8;
 
 module.exports = function dataMarko(props, key, componentDef) {
   var result = "";